Many new features. Read description.
Configuration file (config.ini) extended with more features:
-changing colors of scripts
-changing font and size of scripts

On scriptgui.py there are many things to add if you want more cotrol
over the different colors, fonts and size of all the aspects of the
QsciLexerPython.

On tree.py added the instert action when clicking on a folder to add a
new resource of that kind.