Starting over
Stellar was started long time ago by me and other people. After many
months of learning how to deal with Qt and Python we created a
semi-functional version. Unfortunately the code was horrible and almost
imposible to refactor, so I decided to start over and go step by step
until we get a complex IDE.
You can find the old files on a .zip just in case you want to check it
out and understand why I'm starting again.